feat: category button and PO summary
Created the functionality for filtering out the products using the given
filters.

Reflect products to PO Summary when changing the quantity of a product,
and calculate the subtotal, tax, and total amount for the items in the
PO Summary list.

Implemented the delete button for each item to reflect the subtotal,
tax, and total amount once removed.

Remaining Issues:
- Search function not implemented
- Select different shop not implemented
- Send PO to Supplier not implemented
- When using the delete button in the PO Summary, it will not reset the
  quantity of the product to 0 in the Products List.